backend/requirements.txt
-c constraints.txt
-e src/utrechtsciencepark[test]

zope.testrunner

# Add required add-ons
# (Ideally add them in setup.py for utrechtsciencepark)
# collective.easyform
